pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "helmet-cam-trimmer"
version = "0.0.2"
authors = [
  { name="Tom Phillips", email="tom@tomwphillips.co.uk" },
]
description = "A utility to trim video footage around particular timestamps."
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: OS Independent",
]

requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
trim = "trim:main"

[project.urls]
repository = "https://github.com/tomwphillips/helmet-cam-trimmer"

[tool.setuptools]
py-modules = ["trim"]

[tool.isort]
profile = "black"
